Web_app/requirements.txt
asgiref==3.3.1
blis==0.2.4
boto3==1.17.13
botocore==1.20.13
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cymem==2.0.5
Django==3.1.7
djangorestframework==3.12.2
httpie==2.4.0
idna==2.10
jmespath==0.10.0
joblib==1.0.1
jsonschema==2.6.0
murmurhash==1.0.5
neuralcoref==4.0
numpy==1.20.1
Pillow==8.1.0
plac==0.9.6
preshed==2.0.1
Pygments==2.8.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.3.4
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
sentencepiece==0.1.95
six==1.15.0
sklearn==0.0
spacy==2.1.3
sqlparse==0.4.1
srsly==1.0.5
thinc==7.0.8
threadpoolctl==2.1.0
torch==1.7.1+cu110
torchaudio==0.7.2
torchvision==0.8.2+cu110
tqdm==4.57.0
transformers==2.2.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wasabi==0.8.2
